Refactor copy content
