+++ src/app/components/responsive-menu/responsive-menu.component.scss
@@ -1,6 +1,3 @@
-$size: 2.5rem;
-
-//@include scale-icon-content();
 .cdk-overlay-backdrop-showing {
     &.responsive-menu-overlay-backdrop {
         background-color: unset;
