+++ src/app/services/util.service.ts
@@ -40,4 +40,27 @@ export class UtilService {
     matches(a: number, b: number): boolean {
         return Number(a) === Number(b);
     }
+
+    clipboardCopy(text: string): void {
+        window.focus();
+        setTimeout(() => {
+            try {
+                // Attempt 1
+                void navigator.clipboard.writeText(text);
+            } catch (err1) {
+                // Attempt 2
+                try {
+                    console.error(err1);
+                    const el = document.createElement('textarea');
+                    el.value = text;
+                    document.body.appendChild(el);
+                    el.select();
+                    document.execCommand('copy');
+                    document.body.removeChild(el);
+                } catch (err2) {
+                    console.error(err2);
+                }
+            }
+        });
+    }
 }
