+++ src/app/services/wallet-events.service.ts
@@ -51,6 +51,6 @@ export class WalletEventsService {
     /** Backup active wallet Mnemonic Phrase to clipboard */
     backupMnemonic = new Subject<{ mnemonic: string; openSnackbar: boolean }>();
 
-    /** User has copied address to clipboard. */
+    /** User has copied account address to clipboard. */
     copiedAddress = new Subject<{ address: string }>();
 }
